requirements.txt
tflearn==0.3.2
numpy==1.16.5
tensorflow==1.3.0
gym==0.9.4
Pillow==4.3.0
psutil==5.6.2
statsmodels==0.10.1